add manager in frontend

+++ database/migrations/2025_01_12_140742_add_feature_in_home_to_categories_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->boolean('feature_in_home')->default(false)->after('name');
+            if (!Schema::hasColumn('categories', 'feature_in_home')) {
+                $table->boolean('feature_in_home')->default(false)->after('name');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->dropColumn('feature_in_home');
+            if (Schema::hasColumn('categories', 'feature_in_home')) {
+                $table->dropColumn('feature_in_home');
+            }
         });
     }
 };
